Refact element renderer
* separate out conversion to JSX from wrapping/meta/role styling
* remove role styling parameter for Figure as no longer needed

Also fix some style errors.
