+++ src/model/json-schema.json
@@ -1768,6 +1768,7 @@
                 "halfwidth",
                 "immersive",
                 "inline",
+                "richLink",
                 "showcase",
                 "supporting",
                 "thumbnail"
