+++ src/web/components/MainMedia.tsx
@@ -4,7 +4,7 @@ import { css, cx } from 'emotion';
 import { until } from '@guardian/src-foundations/mq';
 import { Display } from '@guardian/types';
 
-import { ElementRenderer } from '@root/src/web/lib/ElementRenderer';
+import { renderArticleElement } from '@root/src/web/lib/renderElement';
 import { getZIndex } from '@frontend/web/lib/getZIndex';
 
 const mainMedia = css`
@@ -73,18 +73,18 @@ export const MainMedia: React.FC<{
 			format.display === Display.Immersive ? immersiveWrapper : noGutters,
 		)}
 	>
-		{elements.map((element, index) => (
-			<ElementRenderer
-				element={element}
-				format={format}
-				palette={palette}
-				index={index}
-				hideCaption={hideCaption}
-				adTargeting={adTargeting}
-				starRating={starRating}
-				host={host}
-				isMainMedia={true}
-			/>
-		))}
+		{elements.map((element, index) =>
+			renderArticleElement({
+				format,
+				palette,
+				element,
+				adTargeting,
+				host,
+				index,
+				isMainMedia: true,
+				starRating,
+				hideCaption,
+			}),
+		)}
 	</div>
 );
